added a few arrow utils
